Gunicorn+Flask+Postgres+Docker/requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.3
gunicorn==20.1.0
wheel==0.37.1
pytest==7.1.3
